+++ ChangeLog.md
@@ -1,6 +1,6 @@
 ## Library Versions ##
-* development (unreleased)
-    * Use Bitcoin 0.15.0 with indexd as a transaction index
+* v9.56.0 (2018-09-16)
+    * Use Bitcoin 0.16.2 with indexd as a transaction index
     * allow multisig encoding method as the bytespersigop DoS protection was changed in Bitcoin Core v0.13.0
     * API change (backwards compatible): adds extended_tx_info parameter to create methods
     * Breaking API change. The result of get_unspent_txouts has changed and now looks like this: [{